requirements.txt
django
wheel
django-environ
psycopg2-binary
djangorestframework
wikidata